Implement writer of results Generalize some parameter like TOLERANCE or EPSILON Minor changes on k_box and loader Change evaluar_v1 for evaluar_v2 Other minor changes in order to finish the work
