+++ T1/src/base.py
@@ -9,7 +9,7 @@ import time
 print("Usando OpenCV {} Python {}.{}.{}".format(cv2.__version__, sys.version_info.major, sys.version_info.minor, sys.version_info.micro))
 
 # GLOBAL VARRIABLES
-MAIN_FOLDER = "C:/Users/Daniel/Desktop/Semestre2018-1/multimedia/myCC5213/T1/"
+MAIN_FOLDER = str(os.getcwd()) + "/" #" C:/Users/Daniel/Desktop/Semestre2018-1/multimedia/myCC5213/T1/"
 DATA_FOLDER = "data/"
 MAIN_DATA_FOLDER = MAIN_FOLDER + DATA_FOLDER
 C_FOLDER = "comerciales/"
@@ -20,8 +20,11 @@ KNF_FOLDER = "k_nearest_frames/"
 RESULTS_FOLDER = "results/"
 
 FRAMES_PER_CELL = 10
-K = 10
-
+K = 12
+RESIZE = (320, 180) #(160, 90)
+FPS_RATE = 30
+TOLERANCE = 50
+EPSILON = 4
 
 def show_frame(window_name, image, valorAbsoluto= False, escalarMin0Max255= False):
     if valorAbsoluto:
